Added Sprite to Inventory slot
