--- Assets/Scripts/Inventory/InventorySlotUI.cs
@@ -1,22 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using TMPro;
-using UnityEngine;
-using UnityEngine.UI;
-
-public class InventorySlot : MonoBehaviour
-{
-    public Image icon;
-    public TextMeshProUGUI count;
-    // Start is called before the first frame update
-    void Start()
-    {
-        
-    }
-
-    // Update is called once per frame
-    void Update()
-    {
-        
-    }
-}
